organized the structore
